fix(sonar): 修复可靠性 BUG 与安全热点，压制前端误报
消除空指针/类型转换/中断处理等 Java BUG；charts.ts ReDoS 分段解析；
sonar-project 压制 rpx、Lombok 字段、uni page 与 Vue @click a11y 误报。

+++ services/trade-service/src/main/java/com/aicabinet/trade/service/PurchaseSuggestionService.java
@@ -250,7 +250,7 @@ public class PurchaseSuggestionService {
         }
         double sumX = 0, sumY = 0, sumXY = 0, sumXX = 0;
         for (int i = series.length - n; i < series.length; i++) {
-            double x = i - (series.length - n);
+            double x = (double) i - (series.length - n);
             double y = series[i];
             sumX += x;
             sumY += y;
